requirements.txt
gradio
requests
httpx
pandas
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
langchain-openai>=1.0.0
tavily-python>=0.5.0
ddgs>=6.0.0
trafilatura>=1.12.0
pypdf>=5.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
faster-whisper>=1.0.0
youtube-transcript-api>=0.6.0
datasets>=2.14.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest
beautifulsoup4
fal-client
